Move jcad model to schema package
