+++ .eslintignore
@@ -8,3 +8,4 @@ ui-tests
 **/build/
 jupytercad/
 examples/
+packages/opencascade/build.yml
