+++ packages/base/src/jcadplugin/modelfactory.ts
@@ -1,8 +1,11 @@
 import { DocumentRegistry } from '@jupyterlab/docregistry';
 import { Contents } from '@jupyterlab/services';
 
-import { IAnnotationModel, IJupyterCadDoc } from '../types';
-import { JupyterCadModel } from '../model';
+import {
+  IAnnotationModel,
+  IJupyterCadDoc,
+  JupyterCadModel
+} from '@jupytercad/schema';
 
 /**
  * A Model factory to create new instances of JupyterCadModel.
