+++ packages/base/src/panelview/annotations.tsx
@@ -1,8 +1,7 @@
+import { IAnnotationModel } from '@jupytercad/schema';
 import { PanelWithToolbar, ReactWidget } from '@jupyterlab/ui-components';
-
 import * as React from 'react';
 
-import { IAnnotationModel } from '../types';
 import { Annotation } from '../annotation/view';
 
 interface IProps {
