+++ packages/schema/src/index.ts
@@ -1,13 +1,2 @@
-export * from './_interface/any';
-export * from './_interface/box';
-export * from './_interface/cone';
-export * from './_interface/cut';
-export * from './_interface/cylinder';
-export * from './_interface/extrusion';
-export * from './_interface/fuse';
-export * from './_interface/intersection';
-export * from './_interface/jcad';
-export * from './_interface/placement';
-export * from './_interface/sketch';
-export * from './_interface/sphere';
-export * from './_interface/torus';
+export * from './interfaces';
+export * from './model';
